fix(session): detect replaced device on foreground
